se está trabajando en el formulario para añadir un nuevo encuentro
