+++ BichoPedia/src/main/java/com/salesianostriana/dam/bichopedia/controller/UsuarioController.java
@@ -1,7 +1,5 @@
 package com.salesianostriana.dam.bichopedia.controller;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
