+++ BichoPedia/src/main/java/com/salesianostriana/dam/bichopedia/model/Encuentro.java
@@ -40,5 +40,10 @@ public class Encuentro {
 	private double peso;
 	private char sexo;
 	
+	public Encuentro getEncuentro() {
+		
+		return this;
+	}
+	
 	
 }
