--- BichoPedia/src/main/resources/templates/register.html
@@ -1,102 +1,0 @@
-<!DOCTYPE html>
-<html lang="es" xmlns:th="http://www.thymeleaf.org">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-        <link th:href="@{/css/estilos.css}" rel="stylesheet">
-        <link rel="stylesheet" href="../../static/css/estilos.css">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"></script>
-
-    <title>Sing up</title>
-</head>
-
-<body>
-    <header>
-        <nav th:replace="~{fragmentos/menu::menu}" >
-        </nav>
-    </header>
-    <main>
-        <section class="vh-100 login">
-            <div class="container py-5 h-100">
-                <div class="row d-flex justify-content-center align-items-center h-100">
-                    <div class="col-12 col-md-10 col-lg-8 col-xl-6">
-                        <form class=" logincard p-5 text-center" method="POST" th:action="registerOk"
-                            th:object="${registerForm}">
-                            <h2 class="mb-5">Sing up</h2>
-                            <div class="form-outline mb-4 d-flex justify-content-around">
-                                <div class="me-2">
-                                    <div>
-                                        <i class="bi bi-person-circle"></i>
-                                    </div>
-                                    <input type="text" id="nombre" class="form-control form-control-lg"
-                                        placeholder="Nombre" required th:field="*{nombre}" />
-                                    <label class="form-label" for="nombre"></label>
-                                </div>
-                                <div>
-                                    <input type="text" id="apellidos" class="form-control form-control-lg"
-                                        placeholder="Apellidos" required th:field="*{apellidos}" />
-                                    <label class="form-label" for="apellidos"></label>
-                                </div>
-                            </div>
-                            <div class="form-outline mb-4">
-                                <div>
-                                    <div>
-                                        <i class="bi bi-envelope-at"></i>
-                                    </div>
-                                    <input type="email" id="correo" class="form-control form-control-lg"
-                                        placeholder="correo" required th:field="*{correo}" />
-                                    <label class="form-label" for="correo"></label>
-                                </div>
-                            </div>
-                            <div class="form-outline mb-4">
-                                <div>
-                                    <div>
-                                        <i class="bi bi-person-circle"></i>
-                                    </div>
-                                    <input type="text" id="username" class="form-control form-control-lg"
-                                        placeholder="Usuario" required th:field="*{username}" />
-                                    <label class="form-label" for="username"></label>
-                                </div>
-                            </div>
-                            <div class="form-outline mb-4">
-                                <div>
-                                    <div>
-                                        <i class="bi bi-lock"></i>
-                                    </div>
-                                    <input type="password" id="passwd" class="form-control form-control-lg"
-                                        placeholder="Constraseña" required th:field="*{passwd}" />
-                                    <label class="form-label" for="passwd"></label>
-                                </div>
-                            </div>
-                            <div class="form-outline mb-4">
-                                <div>
-                                    <div>
-                                        <i class="bi bi-calendar-event"></i>
-                                    </div>
-                                    <input type="date" id="fechaNac" class="form-control form-control-lg" required />
-                                    <label class="form-label" for="fechaNac"
-                                        th:field="${registerForm.fechaNac}"></label>
-                                </div>
-                            </div>
-                            <div class="d-flex justify-content-around">
-                                <!-- Checkbox -->
-                                <div class="form-check d-flex justify-content-start mb-4">
-                                    <input class="form-check-input me-2" type="checkbox" value="" id="form1Example3" />
-                                    <label class="form-check-label" for="form1Example3"> Recordar contraseña </label>
-                                </div>
-                                <p><a href="login.html" class="registerhref">Log in</a></p>
-                            </div>
-                            <button class="btn btn-primary btn-lg btn-block loginbutton" type="submit">Login</button>
-                        </form>
-                    </div>
-                </div>
-            </div>
-        </section>
-    </main>
-</body>
-
-</html>
